+++ starkbank/request/__init__.py
@@ -1,0 +1,1 @@
+from .__request import (get, post, patch, put, delete)