+++ tests/sdk/test_corporate_withdrawal.py
@@ -3,9 +3,35 @@ from unittest import TestCase, main
 from datetime import date, timedelta
 from starkbank import CorporateWithdrawal
 from tests.utils.user import exampleProject
-from tests.utils.withdrawal import generateExampleWithdrawalJson
+from tests.utils.withdrawal import generateExampleWithdrawalJson, payment_script
+import time
 
 starkbank.user = exampleProject
+
+
+class TestCorporateResourcesChecks(TestCase):
+
+    def test_success_corporate_balance(self):
+        balance = starkbank.corporatebalance.get()
+        initial_balance = balance.amount
+
+        if balance.amount < 0:
+            request = payment_script(balance.amount)
+            i = 0
+            while i <= 5:
+                br_code = starkbank.brcodepayment.get(request[0].id)
+                if br_code.status == "success" or br_code.status == "failed":
+                    time.sleep(5)
+                    break
+                time.sleep(10)
+                i += 1
+        balance = starkbank.corporatebalance.get()
+        if balance.amount == initial_balance:
+            time.sleep(5)
+        balance = starkbank.corporatebalance.get()
+        print("initial_balance:" + str(initial_balance))
+        print("final_balance:" + str(balance.amount))
+        print("diff:" + str(initial_balance - balance.amount))
 
 
 class TestCorporateWithdrawalQuery(TestCase):
@@ -50,7 +76,7 @@ class TestCorporateWithdrawalPost(TestCase):
         example_withdrawal = generateExampleWithdrawalJson()
         withdrawal = starkbank.corporatewithdrawal.create(
             withdrawal=CorporateWithdrawal(
-                amount=example_withdrawal.amount,
+                amount=100,
                 external_id=example_withdrawal.external_id
             )
         )